backend/pyproject.toml
[project]
name = "dozhim-backend"
version = "0.1.0"
description = "Dozhim backend API (FastAPI MVP scaffold)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.13.3",
  "psycopg[binary]>=3.2.3",
  "openpyxl>=3.1.5"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["app"]
testpaths = ["tests"]
